fix: Move `AccountDetails` into separate component and `slots`
