+++ docs/pages/toolpad/core/api/account-details.json
@@ -1,55 +1,29 @@
 {
-  "props": {},
+  "props": {
+    "slotProps": { "type": { "name": "shape", "description": "{ avatar: object }" } },
+    "slots": {
+      "type": { "name": "shape", "description": "{ avatar?: elementType }" },
+      "additionalInfo": { "slotsApi": true }
+    }
+  },
   "name": "AccountDetails",
-  "imports": ["import { AccountDetails } from '@toolpad/core/Account';"],
-  "classes": [
+  "imports": [
+    "import { AccountDetails } from '@toolpad/core/AccountDetails';",
+    "import { AccountDetails } from '@toolpad/core';"
+  ],
+  "slots": [
     {
-      "key": "circular",
-      "className": "",
-      "description": "Styles applied to the root element if `variant=\"circular\"`.",
-      "isGlobal": false
-    },
-    {
-      "key": "colorDefault",
-      "className": "",
-      "description": "Styles applied to the root element if not `src` or `srcSet`.",
-      "isGlobal": false
-    },
-    {
-      "key": "fallback",
-      "className": "",
-      "description": "Styles applied to the fallback icon",
-      "isGlobal": false
-    },
-    {
-      "key": "img",
-      "className": "",
-      "description": "Styles applied to the img element if either `src` or `srcSet` is defined.",
-      "isGlobal": false
-    },
-    {
-      "key": "root",
-      "className": "",
-      "description": "Styles applied to the root element.",
-      "isGlobal": false
-    },
-    {
-      "key": "rounded",
-      "className": "",
-      "description": "Styles applied to the root element if `variant=\"rounded\"`.",
-      "isGlobal": false
-    },
-    {
-      "key": "square",
-      "className": "",
-      "description": "Styles applied to the root element if `variant=\"square\"`.",
-      "isGlobal": false
+      "name": "avatar",
+      "description": "The component used for the Avatar",
+      "default": "Avatar",
+      "class": null
     }
   ],
+  "classes": [],
   "spread": true,
   "themeDefaultProps": false,
   "muiName": "AccountDetails",
-  "filename": "/packages/toolpad-core/src/Account/AccountDetails.tsx",
+  "filename": "/packages/toolpad-core/src/AccountDetails/AccountDetails.tsx",
   "inheritance": null,
   "demos": "<ul><li><a href=\"/toolpad/core/react-account/\">Account</a></li></ul>",
   "cssComponent": false
