+++ docs/translations/api-docs/account-details/account-details.json
@@ -1,33 +1,9 @@
 {
   "componentDescription": "",
-  "propDescriptions": {},
-  "classDescriptions": {
-    "circular": {
-      "description": "Styles applied to {{nodeName}} if {{conditions}}.",
-      "nodeName": "the root element",
-      "conditions": "<code>variant=\"circular\"</code>"
-    },
-    "colorDefault": {
-      "description": "Styles applied to {{nodeName}} if {{conditions}}.",
-      "nodeName": "the root element",
-      "conditions": "not <code>src</code> or <code>srcSet</code>"
-    },
-    "fallback": { "description": "Styles applied to the fallback icon" },
-    "img": {
-      "description": "Styles applied to {{nodeName}} if {{conditions}}.",
-      "nodeName": "the img element",
-      "conditions": "either <code>src</code> or <code>srcSet</code> is defined"
-    },
-    "root": { "description": "Styles applied to the root element." },
-    "rounded": {
-      "description": "Styles applied to {{nodeName}} if {{conditions}}.",
-      "nodeName": "the root element",
-      "conditions": "<code>variant=\"rounded\"</code>"
-    },
-    "square": {
-      "description": "Styles applied to {{nodeName}} if {{conditions}}.",
-      "nodeName": "the root element",
-      "conditions": "<code>variant=\"square\"</code>"
-    }
-  }
+  "propDescriptions": {
+    "slotProps": { "description": "The props used for each slot inside." },
+    "slots": { "description": "The components used for each slot inside." }
+  },
+  "classDescriptions": {},
+  "slotDescriptions": { "avatar": "The component used for the Avatar" }
 }
