+++ packages/toolpad-core/src/AccountDetails/index.ts
@@ -1,0 +1,1 @@
+export * from './AccountDetails';
