+++ packages/toolpad-core/src/index.ts
@@ -6,6 +6,8 @@ export * from './SignInPage';
 
 export * from './Account';
 
+export * from './AccountDetails';
+
 export * from './PageContainer';
 
 export * from './useActivePage';
